requirements.txt
requests
selenium
webdriver_manager
pandas
openpyxl
python-dotenv